apps/workers/requirements.txt
# Created automatically by Cursor AI (2024-08-24)
# Python Workers Dependencies

# Core Framework
celery==5.3.4
redis==5.0.1

# AI/ML & Processing
openai==1.3.0
anthropic==0.7.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23

# Speech & Audio
whisper==1.1.10
librosa==0.10.1
pydub==0.25.1

# Storage & File Handling
boto3==1.34.0
minio==7.2.0
Pillow==10.1.0

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Event Bus
nats-py==2.3.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
click==8.1.7
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
